requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
idna==2.8
pyfiglet==0.8.post1
requests==2.22.0
soupsieve==1.9.5
termcolor==1.1.0
urllib3==1.25.7
Pillow==8.4.0
pywin32==305
